% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{add_drawn_individuals}
\alias{add_drawn_individuals}
\title{Add drawn individuals}
\usage{
add_drawn_individuals(timeperiod_isd, sampling_gmm, draw_seed = NULL)
}
\arguments{
\item{timeperiod_isd}{an ISD-shaped dataframe. what matters is the number of rows and that it have mass and id columns.}

\item{sampling_gmm}{begin or end, just one}

\item{draw_seed}{For the actual sampling.}
}
\value{
dataframe with body mass (and energy use) for those individuals drawn from the sampling_gmm provided
}
\description{
Adds individuals drawn from a sampling GMM to an ISD-shaped dataframe.
}
