% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{all_core_analyses}
\alias{all_core_analyses}
\title{Run and collect all core analyses}
\usage{
all_core_analyses(
  ts_comp,
  begin_years = NULL,
  end_years = NULL,
  isd_seed = NULL
)
}
\arguments{
\item{ts_comp}{a MATSS_style dataset}

\item{begin_years}{optional}

\item{end_years}{optional}

\item{isd_seed}{optional}
}
\value{
results
}
\description{
Run and collect all core analyses
}
