% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{be_comparison}
\alias{be_comparison}
\title{Compare ISD, sp composition}
\usage{
be_comparison(
  dat,
  initial_isd_seed = 1989,
  shuffle_seed = 1989,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  nshuffles = 500,
  n_isd_draws = 5
)
}
\arguments{
\item{dat}{matss}

\item{initial_isd_seed}{1989}

\item{shuffle_seed}{1989}

\item{begin_years}{usual}

\item{end_years}{usual}

\item{nshuffles}{n for ks test}

\item{n_isd_draws}{n for gmms}
}
\value{
dataframe
}
\description{
Compare ISD, sp composition
}
