% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{compare_actual_null_models}
\alias{compare_actual_null_models}
\title{Compare null models to reality}
\usage{
compare_actual_null_models(actual_summary, null_summary)
}
\arguments{
\item{actual_summary}{actual}

\item{null_summary}{null}
}
\value{
df
}
\description{
Compare null models to reality
}
