% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{compare_both_brms}
\alias{compare_both_brms}
\title{Pull comparisons for two sets of brms}
\usage{
compare_both_brms(some_brms_fits)
}
\arguments{
\item{some_brms_fits}{a list}
}
\value{
a dataframe
}
\description{
Pull comparisons for two sets of brms
}
