% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{compare_both_stanarms}
\alias{compare_both_stanarms}
\title{COmpare all fits}
\usage{
compare_both_stanarms(some_stanlms_fits)
}
\arguments{
\item{some_stanlms_fits}{fits}
}
\value{
df
}
\description{
COmpare all fits
}
