% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{compare_brms}
\alias{compare_brms}
\title{LOO compare on brms}
\usage{
compare_brms(brms_fits)
}
\arguments{
\item{brms_fits}{for one currency}
}
\value{
a dataframe of results from loo_compare
}
\description{
LOO compare on brms
}
