% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{compare_isds}
\alias{compare_isds}
\title{Compute ISD turnover}
\usage{
compare_isds(ts_isds, begin_years = NULL, end_years = NULL)
}
\arguments{
\item{ts_isds}{result of simulate_isd_ts}

\item{begin_years}{optional}

\item{end_years}{optional}
}
\value{
df of turnover
}
\description{
Compute ISD turnover
}
