% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{compare_species_composition}
\alias{compare_species_composition}
\title{Compare species composition}
\usage{
compare_species_composition(ts_comp, begin_years = NULL, end_years = NULL)
}
\arguments{
\item{ts_comp}{MATSS dataset}

\item{begin_years}{optional}

\item{end_years}{optional}
}
\value{
dataframe
}
\description{
Compare species composition
}
