% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{compute_raw_sv_change}
\alias{compute_raw_sv_change}
\title{Compute raw change from the first five to the last five years}
\usage{
compute_raw_sv_change(caps_svs)
}
\arguments{
\item{caps_svs}{result of pull_caps(get_annual_svs), must include timeperiod as \code{timeperiod} and cols \code{abundance}, \code{biomass}, \code{energy}, \code{mean_biomass}, \code{mean_energy}}
}
\value{
dataframe of \code{response} and \verb{mean value}
}
\description{
Compute raw change from the first five to the last five years
}
