% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{construct_sampling_gmm}
\alias{construct_sampling_gmm}
\title{Construct GMMs for sampling}
\usage{
construct_sampling_gmm(
  ts_comp,
  n_isd_draws = 5,
  initial_isd_seed = NULL,
  begin_years = 1988:1992,
  end_years = 2014:2018
)
}
\arguments{
\item{ts_comp}{matss-shaped dataset}

\item{n_isd_draws}{how many versions of the ISD to use to construct the density function. There isn't too much variation between them, defaults to 5. Adding more increases time nonlinear due to the GMM, do not recommend much more than 10.}

\item{initial_isd_seed}{don't recommend providing unless testing}

\item{begin_years}{defaults 1988-1992}

\item{end_years}{defaults 2014-2018}
}
\value{
list of two gmm densities as dataframes with columns mass, density, timeperiod
}
\description{
Constructs well-resolved smooths for the begin and end ISDs to use for resampling. Repeatedly draws ISDs by sampling from species' intraspecific body size distributions and combines these repeated samples in to one large sample. Then fits a GMM to this sample to get a very smooth version of the GMM-ified ISD. This allows us to resample ISDs with noninteger ratios of species abundances in a consistent way.
}
