% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continentall_null_model.R
\name{continental_null_model_wrapper}
\alias{continental_null_model_wrapper}
\title{Wrapper for continental null model}
\usage{
continental_null_model_wrapper(
  ts_dat,
  nsims = 100,
  begin_years = NULL,
  end_years = NULL
)
}
\arguments{
\item{ts_dat}{matss dataset}

\item{nsims}{default 100}

\item{begin_years}{pass}

\item{end_years}{pass}
}
\value{
df
}
\description{
Wrapper for continental null model
}
