% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_transient.R
\name{core_transient}
\alias{core_transient}
\title{Remove transient and/or non-core species}
\usage{
core_transient(dat, core_only = F)
}
\arguments{
\item{dat}{matss-style}

\item{core_only}{if T, returns only core (present in > 2/3 of samples). if F, removes transients (present in < 1/3 of samples).}
}
\value{
matss-style but filtered
}
\description{
Remove transient and/or non-core species
}
