% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_transient.R
\name{core_transient_null}
\alias{core_transient_null}
\title{Remove transient and/or non-core species}
\usage{
core_transient_null(dat, core_only = F, null_seed = 1989)
}
\arguments{
\item{dat}{matss-style}

\item{core_only}{if T, returns only core (present in > 2/3 of samples). if F, removes transients (present in < 1/3 of samples).}

\item{null_seed}{1989}
}
\value{
matss-style but filtered
}
\description{
Remove transient and/or non-core species
}
