% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{draw_communities}
\alias{draw_communities}
\title{Draw communities}
\usage{
draw_communities(
  ts_comp,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  draw_seed = NULL,
  sampling_gmms = NULL,
  initial_isd_seed = NULL,
  raw_isd_seed = NULL
)
}
\arguments{
\item{ts_comp}{matss-style dataset}

\item{begin_years}{default 1988:1992.}

\item{end_years}{default 2014:2018.}

\item{draw_seed}{don't provide unless you are doing something specific}

\item{sampling_gmms}{do provide for speed! result of construct_sampling_gmms}

\item{initial_isd_seed}{don't provide unless you have a specific reason to}

\item{raw_isd_seed}{don't provide unless you have a reason to}
}
\value{

}
\description{
High-level fxn to get yearly estimates for the state variables under the scenarios of 1) the ISD is conserved begin-end and 2) the ISD changes begin-end. A way to partition change in biomass or energy use expected given changes in abundance and sampling variation, and changes that must be partially driven by changes in the size shift.
}
