% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{draw_communities_wrapper}
\alias{draw_communities_wrapper}
\title{Wrapper for draw_communities}
\usage{
draw_communities_wrapper(
  ts_comp,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  ndraws = 100,
  draw_seed = NULL,
  sampling_gmms = NULL,
  initial_isd_seed = NULL,
  raw_isd_seed = NULL
)
}
\arguments{
\item{ts_comp}{matss-style dataset}

\item{begin_years}{default 1988:1992.}

\item{end_years}{default 2014:2018.}

\item{ndraws}{default 100}

\item{draw_seed}{don't provide unless you are doing something specific}

\item{sampling_gmms}{do provide for speed! result of construct_sampling_gmms}

\item{initial_isd_seed}{don't provide unless you have a specific reason to}

\item{raw_isd_seed}{don't provide unless you have a reason to}
}
\value{
dataframe of sims
}
\description{
To facilitate pipelines. Just runs draw_communities \code{ndraws} times and tacks on a column keeping track of which sim.
}
