% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{draw_individuals}
\alias{draw_individuals}
\title{Draw individuals from a GMM}
\usage{
draw_individuals(nind, sampling_gmm, draw_seed = NULL)
}
\arguments{
\item{nind}{how many to draw}

\item{sampling_gmm}{a dataframe with mass (on log scale) and density columns.}

\item{draw_seed}{don't recommmend providing}
}
\value{
drawn (log) body masses, and metainfo for whether the sampling gmm is from begin or end and the seed used when drawing
}
\description{
Draw individual body masses from a GMM density function for the ISD
}
