% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{draw_wrapper}
\alias{draw_wrapper}
\title{Tiny wrapper to just get qis}
\usage{
draw_wrapper(winners, fits)
}
\arguments{
\item{winners}{result of loo_select}

\item{fits}{result of fit_stanarms}
}
\value{
qis for draws from winning model
}
\description{
Tiny wrapper to just get qis
}
