% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{draws_df}
\alias{draws_df}
\title{Get draws as a dataframe}
\usage{
draws_df(some_draws)
}
\arguments{
\item{some_draws}{from get_brm_draws}
}
\value{
some_draws as a df
}
\description{
Get draws as a dataframe
}
