% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginend.R
\name{estimate_b}
\alias{estimate_b}
\title{Estimate b (energy)}
\usage{
estimate_b(body_mass)
}
\arguments{
\item{body_mass}{mass}
}
\value{
estimated b (metabolic rate) using pars from Fristoe 2015
}
\description{
Estimate b (energy)
}
