% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{extract_brm_ests}
\alias{extract_brm_ests}
\title{Extract estimates from brms posteriors}
\usage{
extract_brm_ests(some_brms)
}
\arguments{
\item{some_brms}{list with brms te_brm, tb_brm, and matssname. Result of running \code{fit_brms()}}
}
\value{
dataframe with estimates. each row is a draw from the posterior. draws from the 2 models are stacked, group by \code{currency} to get each model.
}
\description{
Extract estimates for a bunch of quantities of interest from the posterior of fitted brms.
}
