% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{extract_diagnostics}
\alias{extract_diagnostics}
\title{Extract model diagnostics}
\usage{
extract_diagnostics(some_fits)
}
\arguments{
\item{some_fits}{some_fits}
}
\value{
data frame of rhats, neffs, divergent
}
\description{
for many unsupervised stanlms
}
