% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{extract_ests}
\alias{extract_ests}
\title{Extract estimates from one brm}
\usage{
extract_ests(a_brm, brm_currency = NULL, matssname = NULL)
}
\arguments{
\item{a_brm}{one of the currency brms}

\item{brm_currency}{which currency "energy" or "biomass"}

\item{matssname}{dataset name}
}
\value{
dataframe with estimates, each row is a draw from the posterior
}
\description{
This is the workhorse function for extracting estimates from the posterior and using them to calculate various quantities.
}
