% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{fit_all_caps_lms}
\alias{fit_all_caps_lms}
\title{Fit linear models to the 5 state variables for the first and last 5 years ("caps")}
\usage{
fit_all_caps_lms(caps_svs)
}
\arguments{
\item{caps_svs}{result of pull_caps(get_annual_svs), must include timeperiod as \code{timeperiod} and cols \code{abundance}, \code{biomass}, \code{energy}, \code{mean_biomass}, \code{mean_energy}}
}
\value{
dataframe of lm results
}
\description{
Fit linear models to the 5 state variables for the first and last 5 years ("caps")
}
