% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{fit_all_timeseries_lms}
\alias{fit_all_timeseries_lms}
\title{Fit linear models to the 5 state variables for the full timeseries}
\usage{
fit_all_timeseries_lms(ts_svs)
}
\arguments{
\item{ts_svs}{dataframe, must contain columns \code{year}, \code{abundance}, \code{biomass}, \code{energy}, \code{mean_biomass}, \code{mean_energy}}
}
\value{
dataframe of lm results
}
\description{
Fit linear models to the 5 state variables for the full timeseries
}
