% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{fit_brms}
\alias{fit_brms}
\title{Fit multiple candidate brms}
\usage{
fit_brms(some_sims, cores = 1, iter = 8000, thin = 2)
}
\arguments{
\item{some_sims}{dataframe with columns matssname, timeperiod, source, total_energy, total_biomass}

\item{cores}{how many cores to use. if on hpg, use ONE. if local, do what you want.}

\item{iter}{how many iterations. at scale, I've been using 4000 to be generous.}
}
\value{
list of brm fit on total_energy, on total_biomass, and dataset name
}
\description{
Fit multiple candidate brms
}
