% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{fit_caps_lm}
\alias{fit_caps_lm}
\title{Fit a lm to the first and last 5 years ("caps")}
\usage{
fit_caps_lm(caps_svs, response_name)
}
\arguments{
\item{caps_svs}{result of pull_caps(get_annual_svs), must include timeperiod as \code{timeperiod}}

\item{response_name}{the name of the variable to compare to year, must be the name of a column in caps_svs}
}
\value{
data frame of response_name, p-value, r2, slope, and fitted ratio (last value / first fitted value) for \code{lm(response_name ~ timeperiod)}
}
\description{
Fit a lm to the first and last 5 years ("caps")
}
