% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{fit_timeseries_lm}
\alias{fit_timeseries_lm}
\title{Fit a lm to the full timeseries}
\usage{
fit_timeseries_lm(ts_svs, response_name)
}
\arguments{
\item{ts_svs}{result of get_annual_svs, must include time as \code{year}}

\item{response_name}{the name of the variable to compare to year, must be the name of a column in ts_svs}
}
\value{
data frame of response_name, p-value, r2, slope, and fitted ratio (last value / first fitted value) for \code{lm(response_name ~ year)}
}
\description{
Fit a lm to the full timeseries
}
