% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{get_annual_svs}
\alias{get_annual_svs}
\title{Calculate state variables for every year given ISDs}
\usage{
get_annual_svs(ts_isds)
}
\arguments{
\item{ts_isds}{result of BBSsize::simulate_isd_ts()}
}
\value{
data frame of state variables for each year, route region info, and seed used when simulating the ISD (if one was)
}
\description{
Calculate state variables for every year given ISDs
}
