% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginend.R
\name{get_begin_end_smooths}
\alias{get_begin_end_smooths}
\title{Get GMMs or KDEs fit to first and last 5 years of data}
\usage{
get_begin_end_smooths(begin_end_isds, smooth_method = "gmm")
}
\arguments{
\item{smooth_method}{"gmm" or "kde", currently only implemented for "gmm"}

\item{dataset}{from BBSsize/MATSS}
}
\value{
dataframe of mass, density start, density end, and density change, plus route identifying info
}
\description{
Get GMMs or KDEs fit to first and last 5 years of data
}
