% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginend.R
\name{get_begin_end_species}
\alias{get_begin_end_species}
\title{Get compositional raw}
\usage{
get_begin_end_species(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
df
}
\description{
Get compositional raw
}
