% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{get_brm_draws}
\alias{get_brm_draws}
\title{Extract draws from brms}
\usage{
get_brm_draws(some_brms)
}
\arguments{
\item{some_brms}{list with $te_brm, $tb_brm, $matssname}
}
\value{
list with $te_draws, $tb_draws, $matssname
}
\description{
Extract draws from brms
}
