% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{get_draw_qis}
\alias{get_draw_qis}
\title{Get QIs for draws}
\usage{
get_draw_qis(some_draws)
}
\arguments{
\item{some_draws}{list with te_draws, tb_draws, matssname}
}
\value{
dataframe of qis
}
\description{
Get QIs for draws
}
