% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{gmm_halfsat}
\alias{gmm_halfsat}
\title{half-way point of a gmm}
\usage{
gmm_halfsat(a_gmm)
}
\arguments{
\item{a_gmm}{a gmm}
}
\value{
halfsat
}
\description{
I think this is just getting you to the median mass, in a v roundabout way. I was interested in it because I'm interested in like 50\% of the community-level biomass/energy use, but this is not that, because the ISD GMMS are fit on individual frequency and not wieghted by individual size. Sidebar.
}
