% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{just_isd}
\alias{just_isd}
\title{Get an ISD}
\usage{
just_isd(ts_comp, isd_seed = NULL)
}
\arguments{
\item{ts_comp}{MATSS-shaped (list with $abundance, $covariates, $metadata)}

\item{isd_seed}{can provide, to standardize results (e.g. for testing). if not provided, one will be drawn randomly and kept track of so you can replicate later.}
}
\value{
dataframe of isd
}
\description{
Tiny wrapper for \code{BBSsize::simulate_isd_ts} to get just the ISD component of that output.
}
