% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_transient.R
\name{just_transient}
\alias{just_transient}
\title{Get only transient or non-core speices}
\usage{
just_transient(dat, transient_only = T)
}
\arguments{
\item{dat}{matss-style}

\item{transient_only}{if T, returns only transient (present in <1/3 of samples). if F, removes core (present in > 2/3 of samples).}
}
\value{
matss-style but filtered
}
\description{
Get only transient or non-core speices
}
