% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{ks_comparison}
\alias{ks_comparison}
\title{KS comparison bootstrapped}
\usage{
ks_comparison(
  dat,
  initial_isd_seed = 1989,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  shuffle_seed = 1989,
  nshuffles = 500
)
}
\arguments{
\item{dat}{matss}

\item{initial_isd_seed}{to draw isd}

\item{begin_years}{usual}

\item{end_years}{usual}

\item{shuffle_seed}{for shuffles}

\item{nshuffles}{how many bootstraps}
}
\value{
df
}
\description{
KS comparison bootstrapped
}
