% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{local_null_model}
\alias{local_null_model}
\title{Local null model}
\usage{
local_null_model(
  ts_dat,
  null_mod_seed = 1989,
  sim_index = 1,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  isd_seed = 1989,
  n_isd_draws = 10,
  ndraws = 100,
  return_actual = FALSE
)
}
\arguments{
\item{ts_dat}{matssdat}

\item{null_mod_seed}{seed to control reshuffling of species}

\item{sim_index}{for wrapping}

\item{begin_years}{default 1988:1992}

\item{end_years}{default 2014:2018}

\item{isd_seed}{default 1989, passed to ssims_wrapper}

\item{n_isd_draws}{default 10, passed to ssims_wrapper}

\item{ndraws}{default 100, passed to ssims_wraper}

\item{return_actual}{return actual (unmodified) results?}
}
\value{
df of sims
}
\description{
Reassign species IDs using all species ever observed on a single route.
}
