% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{lower_quantile}
\alias{lower_quantile}
\title{Lower cutoff for a vector}
\usage{
lower_quantile(vector, lower_cutoff = 0.025)
}
\arguments{
\item{vector}{vector of values}

\item{lower_cutoff}{defaults 0.025}
}
\value{
cutoff value
}
\description{
Lower cutoff for a vector
}
