% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_actual_sims}
\alias{make_actual_sims}
\title{Draw sims according to actual dynamics}
\usage{
make_actual_sims(dat)
}
\arguments{
\item{dat}{matss style}
}
\value{
sims drawn as normal
}
\description{
Draws both actual and sim sims.
}
