% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_nochange_sims}
\alias{make_nochange_sims}
\title{Simulate dynamics repeating begin dynamics for end}
\usage{
make_nochange_sims(dat)
}
\arguments{
\item{dat}{matss-style}
}
\value{
sims drawn as if begin years were also the end years
}
\description{
Simulate dynamics repeating begin dynamics for end
}
