% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_nosizechange_sims}
\alias{make_nosizechange_sims}
\title{Simulate dynamics repeating begin size structure for end}
\usage{
make_nosizechange_sims(dat)
}
\arguments{
\item{dat}{matss style}
}
\value{
sims drawn using begin ss for end (both actual and sim)
}
\description{
Simulate dynamics repeating begin size structure for end
}
