% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{overlap_comparison}
\alias{overlap_comparison}
\title{Compare isd overlap}
\usage{
overlap_comparison(
  dat,
  initial_isd_seed = 1989,
  n_isd_draws = 5,
  begin_years = 1988:1992,
  end_years = 2014:2018
)
}
\arguments{
\item{dat}{mats}

\item{initial_isd_seed}{passed to sampling gmm}

\item{n_isd_draws}{passed to sampling gmm}

\item{begin_years}{ditto}

\item{end_years}{ditto}
}
\value{
df
}
\description{
Compare isd overlap
}
