% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nulls.R
\name{percentile_score}
\alias{percentile_score}
\title{Calculate percentile score}
\usage{
percentile_score(actual_value, null_distribution)
}
\arguments{
\item{actual_value}{value}

\item{null_distribution}{distribution to calculate relative to}
}
\value{
percentile
}
\description{
Calculate percentile score
}
