% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginend.R
\name{process_begin_end_isds}
\alias{process_begin_end_isds}
\title{Get smooths and svs for a pair of isds}
\usage{
process_begin_end_isds(begin_end_isds)
}
\arguments{
\item{begin_end_isds}{list}
}
\value{
list of dfs for sv, smooths, metadata
}
\description{
Get smooths and svs for a pair of isds
}
