% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{pull_caps}
\alias{pull_caps}
\title{Pull first and last 5 years}
\usage{
pull_caps(ts_svs, begin_years = NULL, end_years = NULL)
}
\arguments{
\item{ts_svs}{dataframe, must have col \code{year}}

\item{begin_years}{vector of begin years, or will use first 5 years of ts}

\item{end_years}{vector of end years, or will use last 5 years of ts}
}
\value{
ts filtered to years in begin or end and with column \code{timeperiod} for "begin", "end"
}
\description{
Pull first and last 5 years
}
