% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_null_model.R
\name{regional_null_model}
\alias{regional_null_model}
\title{Regional null model}
\usage{
regional_null_model(
  ts_dat,
  ranges_dat = NULL,
  null_mod_seed = NULL,
  sim_index = NULL,
  begin_years = NULL,
  end_years = NULL,
  isd_seed = NULL
)
}
\arguments{
\item{ts_dat}{a MATSS-shaped dataset}

\item{ranges_dat}{a dataframe of route & species encountered each route}

\item{null_mod_seed}{seed to use when reshuffling species. if none given will be drawn.}

\item{sim_index}{sim index}

\item{begin_years}{to pass to all_core_analyses}

\item{end_years}{to pass to all_core_analyses}

\item{isd_seed}{to pass to all_core_analyses}
}
\value{
results of all_core_analyses
}
\description{
Reassign species IDs using all species whose ranges overlap a route, or were ever observed on that route.
}
