% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_null_model.R
\name{regional_null_model_wrapper}
\alias{regional_null_model_wrapper}
\title{Wrapper for regional null model}
\usage{
regional_null_model_wrapper(
  ts_dat,
  ranges_dat = NULL,
  nsims = 100,
  begin_years = NULL,
  end_years = NULL
)
}
\arguments{
\item{ts_dat}{matss dataset}

\item{ranges_dat}{ranges dat}

\item{nsims}{default 100}

\item{begin_years}{pass}

\item{end_years}{pass}
}
\value{
df
}
\description{
Wrapper for regional null model
}
