% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nulls.R
\name{ses}
\alias{ses}
\title{Calculate standardized effect size}
\usage{
ses(actual_value, null_distribution)
}
\arguments{
\item{actual_value}{value}

\item{null_distribution}{distribution to calculate relative to}
}
\value{
ses
}
\description{
Calculate standardized effect size
}
