% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continentall_null_model.R
\name{shuffle_continental}
\alias{shuffle_continental}
\title{Shuffle species for continental null model}
\usage{
shuffle_continental(ts_dat, null_mod_seed)
}
\arguments{
\item{ts_dat}{matss sytle}

\item{null_mod_seed}{seed}
}
\value{
shuffled
}
\description{
Shuffle species for continental null model
}
