% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{shuffle_local}
\alias{shuffle_local}
\title{Perform species shuffling for local null model}
\usage{
shuffle_local(ts_dat, null_mod_seed = NULL)
}
\arguments{
\item{ts_dat}{matss-style}

\item{null_mod_seed}{or null}
}
\value{
matss-style, but shuffled
}
\description{
Perform species shuffling for local null model
}
