% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_null_model.R
\name{shuffle_regional}
\alias{shuffle_regional}
\title{Shuffle species for regional null model}
\usage{
shuffle_regional(ts_dat, ranges_dat, null_mod_seed)
}
\arguments{
\item{ts_dat}{matss-style}

\item{ranges_dat}{ranges}

\item{null_mod_seed}{or null}
}
\value{
matss-style with additional metadata
}
\description{
Shuffle species for regional null model
}
