% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulls.R
\name{shuffle_species}
\alias{shuffle_species}
\title{Reassign species randomly}
\usage{
shuffle_species(dataset, seed = NULL)
}
\arguments{
\item{dataset}{a dataset}

\item{seed}{optional}
}
\value{
modified dataset
}
\description{
Will preserve turnover, SAD, but mess with ISD agnostic to size.
}
