% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_dataset}
\alias{slice_dataset}
\title{Slice a dataset to given start and end}
\usage{
slice_dataset(dataset, startyear, endyear)
}
\arguments{
\item{dataset}{from MATSS}

\item{startyear}{startyear}

\item{endyear}{endyear}
}
\value{
dataset subset to start:end years
}
\description{
Slice a dataset to given start and end
}
