% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{sp_comparison}
\alias{sp_comparison}
\title{Compare species composition begin-end}
\usage{
sp_comparison(dat, begin_years = 1988:1992, end_years = 2014:2018)
}
\arguments{
\item{dat}{matss}

\item{begin_years}{usual}

\item{end_years}{usual}
}
\value{
df
}
\description{
Compare species composition begin-end
}
