% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{ssims_wrapper}
\alias{ssims_wrapper}
\title{Wrapper for getting just summarized sims}
\usage{
ssims_wrapper(
  dat,
  simtype = "actual",
  initial_isd_seed_gmm = 1989,
  n_isd_draws = 10,
  ndraws = 100,
  initial_draw_seed = 1989
)
}
\arguments{
\item{dat}{matss}

\item{simtype}{character}

\item{initial_isd_seed_gmm}{passed to construct_sampling_gmm. seed used to start constructing the isds that are used to fit the gmm.}

\item{n_isd_draws}{default 10}

\item{ndraws}{default 100 but try fewer?}

\item{initial_draw_seed}{seed used to draw 4 seeds which are then used to draw the different time period individuals.}
}
\value{
df
}
\description{
For drake to avoid saving all the sims
}
