% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{summarize_brm_ests}
\alias{summarize_brm_ests}
\title{Summarize brm ests}
\usage{
summarize_brm_ests(some_ests)
}
\arguments{
\item{some_ests}{dataframe with columns matssname, currency, rowindex, and parameters/values of interest. result of extract_brm_ests}
}
\value{
dataframe summarized by matssname and currency to get mean and upper/lower qs.
}
\description{
Summarize draws from the posterior to give mean and upper/lower quantile estimates for quantities of interest (parameters, derived estimates)
}
