% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{summarize_null_models}
\alias{summarize_null_models}
\title{Return ratio of mean e/b begin:end}
\usage{
summarize_null_models(nm_results)
}
\arguments{
\item{nm_results}{result of null_model_wrapper}
}
\value{
df of summary results
}
\description{
Return ratio of mean e/b begin:end
}
