% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nulls.R
\name{summarize_null_results}
\alias{summarize_null_results}
\title{Summarize actual v. null results}
\usage{
summarize_null_results(null_results, actual_results)
}
\arguments{
\item{null_results}{df}

\item{actual_results}{df}
}
\value{
df
}
\description{
Summarize actual v. null results
}
