% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{summarize_sims}
\alias{summarize_sims}
\title{Summarize across sims}
\usage{
summarize_sims(sims)
}
\arguments{
\item{sims}{sims df}
}
\value{
sims df summarized to mean energy and biomass for each year*source
}
\description{
Summarize across sims
}
