% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{upper_quantile}
\alias{upper_quantile}
\title{Upper cutoff for a vector}
\usage{
upper_quantile(vector, upper_cutoff = 0.975)
}
\arguments{
\item{vector}{vector of values}

\item{upper_cutoff}{defaults 0.975}
}
\value{
cutoff value
}
\description{
Upper cutoff for a vector
}
