% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fxns.R
\name{winner_draws}
\alias{winner_draws}
\title{Extract draws from winning models}
\usage{
winner_draws(some_winners, some_models)
}
\arguments{
\item{some_winners}{df with rows for energy and biomass, column "model" name of best model}

\item{some_models}{list of models from fit_brms}
}
\value{
df of draws from both models
}
\description{
Extract draws from winning models
}
