% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{add_gmm}
\alias{add_gmm}
\title{Add GMM density to an ISD}
\usage{
add_gmm(isd)
}
\arguments{
\item{isd}{from BBSsize::simulate_isd_ts}
}
\value{
dataframe with columns mass, density
}
\description{
Fits to log of mass
}
