% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{draw_communities}
\alias{draw_communities}
\title{Draw communities}
\usage{
draw_communities(
  ts_comp,
  begin_years = 1988:1992,
  end_years = 2014:2018,
  draw_seed = 1989,
  sampling_gmms
)
}
\arguments{
\item{ts_comp}{matss-style dataset}

\item{begin_years}{default 1988:1992.}

\item{end_years}{default 2014:2018.}

\item{draw_seed}{This is the seed used to draw the seeds that are then used to draw the individuals for the different sims.}

\item{sampling_gmms}{do provide for speed! result of construct_sampling_gmms}

\item{initial_isd_seed}{If sampling_gmms is not provided, this is the seed that will be used to pull the ISD(s) used to create the sampling gmms. I strongly recommend providing sampling_gmms.}

\item{raw_isd_seed}{This is the seed used to draw the "raw" values for comparison.}
}
\value{
df of sim inds
}
\description{
High-level fxn to get yearly estimates for the state variables under the scenarios of 1) the ISD is conserved begin-end and 2) the ISD changes begin-end. A way to partition change in biomass or energy use expected given changes in abundance and sampling variation, and changes that must be partially driven by changes in the size shift.
}
