% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fxns.R
\name{interaction_lms}
\alias{interaction_lms}
\title{Interaction of sv change}
\usage{
interaction_lms(caps_svs, scaling = "sqrt")
}
\arguments{
\item{caps_svs}{caps_svs}

\item{scaling}{"sqrt" or "rs". defaults to sqrt following Dornelas et al 2019}
}
\value{
df
}
\description{
Interaction of sv change
}
