% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{local_null_model}
\alias{local_null_model}
\title{Local null model}
\usage{
local_null_model(
  ts_dat,
  null_mod_seed = NULL,
  sim_index = NULL,
  begin_years = NULL,
  end_years = NULL,
  isd_seed = NULL
)
}
\arguments{
\item{ts_dat}{a MATSS-shaped dataset}

\item{null_mod_seed}{seed to use when reshuffling species. if none given will be drawn.}

\item{sim_index}{sim index to track}

\item{begin_years}{to pass to all_core_analyses}

\item{end_years}{to pass to all_core_analyses}

\item{isd_seed}{to pass to all_core_analyses}
}
\value{
results of all_core_analyses
}
\description{
Reassign species IDs using all species ever observed on a single route.
}
