% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_null_model.R
\name{local_null_model_wrapper}
\alias{local_null_model_wrapper}
\title{Wrapper for local null model}
\usage{
local_null_model_wrapper(
  ts_dat,
  nsims = 100,
  begin_years = NULL,
  end_years = NULL
)
}
\arguments{
\item{ts_dat}{matss dataset}

\item{nsims}{default 100}

\item{begin_years}{pass}

\item{end_years}{pass}
}
\value{
df
}
\description{
Wrapper for local null model
}
