% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_actual_sims}
\alias{make_actual_sims}
\title{Draw sims according to actual dynamics}
\usage{
make_actual_sims(
  dat,
  initial_isd_seed_gmm = 1989,
  n_isd_draws = 10,
  ndraws = 100,
  initial_draw_seed = NULL
)
}
\arguments{
\item{dat}{matss style}

\item{initial_isd_seed_gmm}{passed to construct_sampling_gmm, use for reproducibility}

\item{n_isd_draws}{default 10}

\item{ndraws}{default 100 but try fewer?}

\item{initial_draw_seed}{ditto}
}
\value{
sims drawn as normal
}
\description{
Draws both actual and sim sims.
}
