% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_nochange_sims}
\alias{make_nochange_sims}
\title{Simulate dynamics repeating begin dynamics for end}
\usage{
make_nochange_sims(
  dat,
  initial_isd_seed_gmm = 1989,
  n_isd_draws = 10,
  ndraws = 100,
  initial_draw_seed = 1989
)
}
\arguments{
\item{dat}{matss-style}

\item{initial_isd_seed_gmm}{passed to construct_sampling_gmm, use for reproducibility}

\item{n_isd_draws}{default 10}

\item{ndraws}{default 100 but try fewer?}

\item{initial_draw_seed}{ditto}
}
\value{
sims drawn as if begin years were also the end years
}
\description{
Simulate dynamics repeating begin dynamics for end
}
