% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{make_nosizechange_sims}
\alias{make_nosizechange_sims}
\title{Simulate dynamics repeating begin size structure for end}
\usage{
make_nosizechange_sims(
  dat,
  initial_isd_seed_gmm = 1989,
  n_isd_draws = 10,
  ndraws = 100,
  initial_draw_seed = 1989
)
}
\arguments{
\item{dat}{matss style}

\item{initial_isd_seed_gmm}{passed to construct_sampling_gmm, use for reproducibility}

\item{n_isd_draws}{default 10}

\item{ndraws}{default 100 but try fewer}

\item{initial_draw_seed}{ditto}
}
\value{
sims drawn using begin ss for end (both actual and sim)
}
\description{
Simulate dynamics repeating begin size structure for end
}
