% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{ssims_wrapper}
\alias{ssims_wrapper}
\title{Wrapper for getting just summarized sims}
\usage{
ssims_wrapper(dat, simtype = "actual")
}
\arguments{
\item{dat}{matss}

\item{simtype}{character}
}
\value{
df
}
\description{
For drake to avoid saving all the sims
}
