% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resim_fxns.R
\name{ssims_wrapper}
\alias{ssims_wrapper}
\title{Wrapper for getting just summarized sims}
\usage{
ssims_wrapper(
  dat,
  simtype = "actual",
  initial_isd_seed_gmm = NULL,
  n_isd_draws = 10,
  ndraws = 100,
  initial_isd_seed_sim = NULL,
  raw_isd_seed = NULL,
  initial_draw_seed = NULL
)
}
\arguments{
\item{dat}{matss}

\item{simtype}{character}

\item{initial_isd_seed_gmm}{passed to construct_sampling_gmm, use for reproducibility}

\item{n_isd_draws}{default 10}

\item{ndraws}{default 100 but try fewer?}

\item{initial_isd_seed_sim}{passed to draw_communities_wrapper, use for reproducibility}

\item{raw_isd_seed}{ditto}

\item{initial_draw_seed}{ditto}
}
\value{
df
}
\description{
For drake to avoid saving all the sims
}
